Dissolve the lib/upgrades module
