+++ .github/CODEOWNERS
@@ -88,7 +88,6 @@
 /client/lib/store-transactions/ @Automattic/payments
 /client/lib/stripe/ @Automattic/payments
 /client/lib/transaction/ @Automattic/payments
-/client/lib/upgrades/ @Automattic/payments
 /client/me/billing-history/ @Automattic/payments
 /client/me/memberships/ @Automattic/payments
 /client/me/pending-payments/ @Automattic/payments
