+++ client/components/data/domain-management/index.jsx
@@ -13,6 +13,7 @@ import { connect } from 'react-redux';
  */
 import CartStore from 'lib/cart/store';
 import DnsStore from 'lib/domains/dns/store';
+import { fetchDns } from 'lib/domains/dns/actions';
 import { fetchUsers } from 'lib/users/actions';
 import { getCurrentUser } from 'state/current-user/selectors';
 import { getPlansBySite } from 'state/sites/plans/selectors';
@@ -20,7 +21,7 @@ import { getSelectedSite } from 'state/ui/selectors';
 import { getDecoratedSiteDomains, isRequestingSiteDomains } from 'state/sites/domains/selectors';
 import { getProductsList } from 'state/products-list/selectors';
 import NameserversStore from 'lib/domains/nameservers/store';
-import { fetchDns, fetchNameservers, fetchWapiDomainInfo } from 'lib/upgrades/actions';
+import { fetchNameservers } from 'lib/domains/nameservers/actions';
 import PageViewTracker from 'lib/analytics/page-view-tracker';
 import QueryContactDetailsCache from 'components/data/query-contact-details-cache';
 import QueryProductsList from 'components/data/query-products-list';
@@ -30,6 +31,7 @@ import SiteRedirectStore from 'lib/domains/site-redirect/store';
 import StoreConnection from 'components/data/store-connection';
 import UsersStore from 'lib/users/store';
 import WapiDomainInfoStore from 'lib/domains/wapi-domain-info/store';
+import { fetchWapiDomainInfo } from 'lib/domains/wapi-domain-info/actions';
 
 function getStateFromStores( props ) {
 	return {
