+++ client/lib/domains/dns/action-types.js
@@ -1,0 +1,10 @@
+export const DNS_ADD = 'DNS_ADD';
+export const DNS_ADD_COMPLETED = 'DNS_ADD_COMPLETED';
+export const DNS_ADD_FAILED = 'DNS_ADD_FAILED';
+export const DNS_APPLY_TEMPLATE_COMPLETED = 'DNS_APPLY_TEMPLATE_COMPLETED';
+export const DNS_DELETE = 'DNS_DELETE';
+export const DNS_DELETE_COMPLETED = 'DNS_DELETE_COMPLETED';
+export const DNS_DELETE_FAILED = 'DNS_DELETE_FAILED';
+export const DNS_FETCH = 'DNS_FETCH';
+export const DNS_FETCH_COMPLETED = 'DNS_FETCH_COMPLETED';
+export const DNS_FETCH_FAILED = 'DNS_FETCH_FAILED';
