+++ client/lib/domains/nameservers/action-types.js
@@ -1,0 +1,4 @@
+export const NAMESERVERS_FETCH = 'NAMESERVERS_FETCH';
+export const NAMESERVERS_FETCH_COMPLETED = 'NAMESERVERS_FETCH_COMPLETED';
+export const NAMESERVERS_FETCH_FAILED = 'NAMESERVERS_FETCH_FAILED';
+export const NAMESERVERS_UPDATE_COMPLETED = 'NAMESERVERS_UPDATE_COMPLETED';
