+++ client/lib/domains/site-redirect/action-types.js
@@ -1,0 +1,7 @@
+export const SITE_REDIRECT_FETCH = 'SITE_REDIRECT_FETCH';
+export const SITE_REDIRECT_FETCH_COMPLETED = 'SITE_REDIRECT_FETCH_COMPLETED';
+export const SITE_REDIRECT_FETCH_FAILED = 'SITE_REDIRECT_FETCH_FAILED';
+export const SITE_REDIRECT_NOTICE_CLOSE = 'SITE_REDIRECT_NOTICE_CLOSE';
+export const SITE_REDIRECT_UPDATE = 'SITE_REDIRECT_UPDATE';
+export const SITE_REDIRECT_UPDATE_COMPLETED = 'SITE_REDIRECT_UPDATE_COMPLETED';
+export const SITE_REDIRECT_UPDATE_FAILED = 'SITE_REDIRECT_UPDATE_FAILED';
