+++ client/lib/domains/wapi-domain-info/action-types.js
@@ -1,0 +1,8 @@
+export const DOMAIN_TRANSFER_ACCEPT_COMPLETED = 'DOMAIN_TRANSFER_ACCEPT_COMPLETED';
+export const DOMAIN_TRANSFER_CANCEL_REQUEST_COMPLETED = 'DOMAIN_TRANSFER_CANCEL_REQUEST_COMPLETED';
+export const DOMAIN_TRANSFER_CODE_REQUEST_COMPLETED = 'DOMAIN_TRANSFER_CODE_REQUEST_COMPLETED';
+export const DOMAIN_TRANSFER_DECLINE_COMPLETED = 'DOMAIN_TRANSFER_DECLINE_COMPLETED';
+export const PRIVACY_PROTECTION_ENABLE_COMPLETED = 'PRIVACY_PROTECTION_ENABLE_COMPLETED';
+export const WAPI_DOMAIN_INFO_FETCH = 'WAPI_DOMAIN_INFO_FETCH';
+export const WAPI_DOMAIN_INFO_FETCH_COMPLETED = 'WAPI_DOMAIN_INFO_FETCH_COMPLETED';
+export const WAPI_DOMAIN_INFO_FETCH_FAILED = 'WAPI_DOMAIN_INFO_FETCH_FAILED';
