+++ client/lib/domains/wapi-domain-info/reducer.js
@@ -18,7 +18,7 @@ import {
 	WAPI_DOMAIN_INFO_FETCH,
 	WAPI_DOMAIN_INFO_FETCH_COMPLETED,
 	WAPI_DOMAIN_INFO_FETCH_FAILED,
-} from 'lib/upgrades/action-types';
+} from './action-types';
 
 const initialDomainState = {
 	hasLoadedFromServer: false,
@@ -72,9 +72,9 @@ function reducer( state, payload ) {
 				} ),
 			} );
 
-		case DOMAIN_TRANSFER_CODE_REQUEST_COMPLETED:
-			const { data } = state[ action.domainName ],
-				locked = ! action.unlock && data.locked;
+		case DOMAIN_TRANSFER_CODE_REQUEST_COMPLETED: {
+			const { data } = state[ action.domainName ];
+			const locked = ! action.unlock && data.locked;
 
 			return updateDomainState( state, action.domainName, {
 				data: Object.assign( {}, state[ action.domainName ].data, {
@@ -82,6 +82,7 @@ function reducer( state, payload ) {
 				} ),
 				needsUpdate: true,
 			} );
+		}
 
 		case DOMAIN_TRANSFER_ACCEPT_COMPLETED:
 		case DOMAIN_TRANSFER_DECLINE_COMPLETED:
