--- client/lib/upgrades/action-types.js
@@ -1,37 +1,0 @@
-/** @format */
-
-export const DNS_ADD = 'DNS_ADD';
-export const DNS_ADD_COMPLETED = 'DNS_ADD_COMPLETED';
-export const DNS_ADD_FAILED = 'DNS_ADD_FAILED';
-export const DNS_APPLY_TEMPLATE_COMPLETED = 'DNS_APPLY_TEMPLATE_COMPLETED';
-export const DNS_DELETE = 'DNS_DELETE';
-export const DNS_DELETE_COMPLETED = 'DNS_DELETE_COMPLETED';
-export const DNS_DELETE_FAILED = 'DNS_DELETE_FAILED';
-export const DNS_FETCH = 'DNS_FETCH';
-export const DNS_FETCH_COMPLETED = 'DNS_FETCH_COMPLETED';
-export const DNS_FETCH_FAILED = 'DNS_FETCH_FAILED';
-export const DOMAIN_TRANSFER_ACCEPT_COMPLETED = 'DOMAIN_TRANSFER_ACCEPT_COMPLETED';
-export const DOMAIN_TRANSFER_CANCEL_REQUEST_COMPLETED = 'DOMAIN_TRANSFER_CANCEL_REQUEST_COMPLETED';
-export const DOMAIN_TRANSFER_CODE_REQUEST_COMPLETED = 'DOMAIN_TRANSFER_CODE_REQUEST_COMPLETED';
-export const DOMAIN_TRANSFER_DECLINE_COMPLETED = 'DOMAIN_TRANSFER_DECLINE_COMPLETED';
-export const DOMAINS_FETCH = 'DOMAINS_FETCH';
-export const DOMAINS_FETCH_COMPLETED = 'DOMAINS_FETCH_COMPLETED';
-export const DOMAINS_FETCH_FAILED = 'DOMAINS_FETCH_FAILED';
-export const DOMAINS_INITIALIZE = 'DOMAINS_INITIALIZE';
-export const ICANN_VERIFICATION_RESEND_COMPLETED = 'ICANN_VERIFICATION_RESEND_COMPLETED';
-export const NAMESERVERS_FETCH = 'NAMESERVERS_FETCH';
-export const NAMESERVERS_FETCH_COMPLETED = 'NAMESERVERS_FETCH_COMPLETED';
-export const NAMESERVERS_FETCH_FAILED = 'NAMESERVERS_FETCH_FAILED';
-export const NAMESERVERS_UPDATE_COMPLETED = 'NAMESERVERS_UPDATE_COMPLETED';
-export const PRIMARY_DOMAIN_SET = 'PRIMARY_DOMAIN_SET';
-export const PRIVACY_PROTECTION_ENABLE_COMPLETED = 'PRIVACY_PROTECTION_ENABLE_COMPLETED';
-export const SITE_REDIRECT_FETCH = 'SITE_REDIRECT_FETCH';
-export const SITE_REDIRECT_FETCH_COMPLETED = 'SITE_REDIRECT_FETCH_COMPLETED';
-export const SITE_REDIRECT_FETCH_FAILED = 'SITE_REDIRECT_FETCH_FAILED';
-export const SITE_REDIRECT_NOTICE_CLOSE = 'SITE_REDIRECT_NOTICE_CLOSE';
-export const SITE_REDIRECT_UPDATE = 'SITE_REDIRECT_UPDATE';
-export const SITE_REDIRECT_UPDATE_COMPLETED = 'SITE_REDIRECT_UPDATE_COMPLETED';
-export const SITE_REDIRECT_UPDATE_FAILED = 'SITE_REDIRECT_UPDATE_FAILED';
-export const WAPI_DOMAIN_INFO_FETCH = 'WAPI_DOMAIN_INFO_FETCH';
-export const WAPI_DOMAIN_INFO_FETCH_COMPLETED = 'WAPI_DOMAIN_INFO_FETCH_COMPLETED';
-export const WAPI_DOMAIN_INFO_FETCH_FAILED = 'WAPI_DOMAIN_INFO_FETCH_FAILED';
