+++ client/me/purchases/manage-purchase/auto-renew-toggle/index.jsx
@@ -12,7 +12,7 @@ import { localize } from 'i18n-calypso';
  * Internal dependencies
  */
 import { isExpiring } from 'lib/purchases';
-import { disableAutoRenew, enableAutoRenew } from 'lib/upgrades/actions';
+import { disableAutoRenew, enableAutoRenew } from 'lib/purchases/actions';
 import { getCurrentUserId } from 'state/current-user/selectors';
 import { isFetchingUserPurchases } from 'state/purchases/selectors';
 import { fetchUserPurchases } from 'state/purchases/actions';
