+++ client/my-sites/checkout/checkout/notices.jsx
