+++ client/my-sites/checkout/checkout/secure-payment-form.jsx
@@ -46,7 +46,7 @@ import isPresalesChatAvailable from 'state/happychat/selectors/is-presales-chat-
 import getCountries from 'state/selectors/get-countries';
 import QueryPaymentCountries from 'components/data/query-countries/payments';
 import { INPUT_VALIDATION, RECEIVED_WPCOM_RESPONSE } from 'lib/store-transactions/step-types';
-import { displayError, clear } from 'lib/upgrades/notices';
+import { displayError, clear } from './notices';
 import { isEbanxCreditCardProcessingEnabledForCountry } from 'lib/checkout/processor-specific';
 import { planHasFeature } from 'lib/plans';
 import { FEATURE_UPLOAD_PLUGINS, FEATURE_UPLOAD_THEMES } from 'lib/plans/constants';
