+++ client/my-sites/domains/domain-management/components/icann-verification/index.jsx
@@ -12,7 +12,7 @@ import { connect } from 'react-redux';
  * Internal dependencies
  */
 import { EMAIL_VALIDATION_AND_VERIFICATION } from 'lib/url/support';
-import { resendIcannVerification } from 'lib/upgrades/actions';
+import { resendIcannVerification } from 'lib/domains';
 import { errorNotice } from 'state/notices/actions';
 import { domainManagementEditContactInfo } from 'my-sites/domains/paths';
 import getRegistrantWhois from 'state/selectors/get-registrant-whois';
