+++ client/my-sites/domains/domain-management/dns/dns-add-new.jsx
@@ -23,7 +23,7 @@ import TxtRecord from './txt-record';
 import SrvRecord from './srv-record';
 import formState from 'lib/form-state';
 import { errorNotice, successNotice } from 'state/notices/actions';
-import { addDns } from 'lib/upgrades/actions';
+import { addDns } from 'lib/domains/dns/actions';
 import { validateAllFields, getNormalizedData } from 'lib/domains/dns';
 
 class DnsAddNew extends React.Component {
