+++ client/my-sites/domains/domain-management/site-redirect/index.jsx
@@ -27,7 +27,7 @@ import {
 	closeSiteRedirectNotice,
 	fetchSiteRedirect,
 	updateSiteRedirect,
-} from 'lib/upgrades/actions';
+} from 'lib/domains/site-redirect/actions';
 import Card from 'components/card/compact';
 import SectionHeader from 'components/section-header';
 import { composeAnalytics, recordGoogleEvent, recordTracksEvent } from 'state/analytics/actions';
