+++ client/my-sites/domains/domain-management/transfer/transfer-out/locked.jsx
@@ -13,7 +13,7 @@ import Card from 'components/card';
 import SectionHeader from 'components/section-header';
 import { getSelectedDomain } from 'lib/domains';
 import Button from 'components/button';
-import { fetchWapiDomainInfo, requestTransferCode } from 'lib/upgrades/actions';
+import { fetchWapiDomainInfo, requestTransferCode } from 'lib/domains/wapi-domain-info/actions';
 import { displayRequestTransferCodeResponseNotice } from './shared';
 import { TRANSFER_DOMAIN_REGISTRATION } from 'lib/url/support';
