+++ client/my-sites/site-settings/disconnect-site/confirm.jsx
@@ -19,7 +19,7 @@ import NavigationLink from 'components/wizard/navigation-link';
 import redirectNonJetpack from 'my-sites/site-settings/redirect-non-jetpack';
 import { getCurrentPlan } from 'state/sites/plans/selectors';
 import { getSelectedSite, getSelectedSiteId, getSelectedSiteSlug } from 'state/ui/selectors';
-import { submitSurvey } from 'lib/upgrades/actions';
+import { submitSurvey } from 'lib/purchases/actions';
 
 class ConfirmDisconnection extends PureComponent {
 	static propTypes = {
